Move Permission enum into seperate file
