+++ src/bot/buttons/ManageSubscriptions.ts
@@ -2,8 +2,8 @@ import { ButtonInteraction, Message, GuildMember, User, EmbedBuilder, ActionRowB
 import Button from "../structs/Button";
 import players from "../../database/schemas/players";
 import { client, colors } from "../bot";
-import { Permission } from "../../libs/RoleManager";
 import { getSkus } from "../../libs/SkuManager";
+import { Permission } from "../../types/Permission";
 
 const skus = getSkus();
