+++ src/bot/buttons/SetPosition.ts
@@ -3,8 +3,8 @@ import Button from "../structs/Button";
 import players from "../../database/schemas/players";
 import { colors } from "../bot";
 import { constantCase, pascalCase } from "change-case";
-import { Permission } from "../../libs/RoleManager";
 import GlobalPosition from "../../types/GlobalPosition";
+import { Permission } from "../../types/Permission";
 
 const positions = Object.keys(GlobalPosition)
     .filter((pos) => isNaN(Number(pos)))
