+++ src/types/Permission.ts
@@ -1,0 +1,12 @@
+export enum Permission {
+    BypassValidation,
+    CustomIcon,
+    ManageBans,
+    ManageNotes,
+    ManageSubscriptions,
+    ManageReports,
+    ManageRoles,
+    ManageTags,
+    ManageWatchlist,
+    ReportImmunity
+}
